Refactor sandbox module cache
