+++ packages/next/server/web/sandbox/index.ts
@@ -1,1 +1,2 @@
 export * from './sandbox'
+export { clearModuleContext } from './context'
